Add support for disabled entries
I want to be able to add disabled entries to the dropdown so that users
are aware the entry exists but are unable to select it.

I have updated the fuzzyDropdownInit function to add information on if
the iterable is disabled or not for use latter. The list generation has
been updated to add the s_unavailable class if the entry is disabled and
the click hander has been updated to check if the field has the
s_unavailable class set. If so that click is simply ignored disabling
the field. The use of a class appears to be the best approach as it
allows for the field to easily be styled differently as different users
want to indicate that the field can't be used.

The html generation method has also been updated to use jQuery objects
and the append() function instead of strings and the html() function.
The html() function allows javascript to be run so become a vector for
javascript injections if users are not careful. Using jQuery objects
doesn't run any javascript so it safer to use.